refactor(acp): move spawn-flag helpers into acp_client/spawn_flags.rs
acp_client.rs is the largest Rust source file. Its contiguous 748-line
block of spawn-side pure helpers — CLI flag builders (*_spawn_flags),
capability probes (cli_supports_*), normalizers, and env mutation — had
no dependency on AcpClient state and already reads as one domain.

Moves the block verbatim into src-tauri/src/acp_client/spawn_flags.rs
(file + sibling dir layout, no mod.rs churn). acp_client.rs re-exports
the module (`pub use spawn_flags::*`) so every `crate::acp_client::*`
call site and `use super::*` test module resolves unchanged; only the
private apply_grok_build_proxy_env became pub(crate) for the parent.

acp_client.rs: 7252 -> 6504 lines.
